pyproject.toml
[project]
name = "reborn"
version = "0.1.0"
description = "Personal AI assistant powered by Codex App Server and Claude SDK"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "claude-agent-sdk>=0.1.40",
    "fastmcp>=2.12.0",
    "fastapi>=0.129.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot[ext]>=22.6",
    "pyyaml>=6.0.3",
    "slack-bolt>=1.27.0",
    "uvicorn>=0.41.0",
    "loguru>=0.7.3",
    "pymupdf>=1.27.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
